requirements.txt
flask==1.1.2
Pillow
pyheif==0.5.1
werkzeug==2.0.2
jinja2==3.0.1
itsdangerous==1.1.0
